refactor(ingester): extract the populateSuggestions function
