+++ .gitlab-ci.yml
@@ -40,3 +40,14 @@ stages:
     include:
       - local: /infra/logger/.gitlab-ci.yml
     strategy: depend
+
+ingester:
+  stage: Code Quality
+  needs:
+    - job: "@socialgouv/cdtn-logger"
+  trigger:
+    include:
+      - local: /modules/ingester/.gitlab-ci.yml
+    strategy: depend
+  variables:
+    PARENT_PIPELINE_ID: $CI_PIPELINE_ID
