+++ modules/ingester/.gitlab-ci.yml
@@ -1,0 +1,136 @@
+#
+#
+#
+
+variables:
+  #
+  PACKAGE: ingester
+  PACKAGE_CWD: ${CI_PROJECT_DIR}/modules/${PACKAGE}
+  #
+  CACHE_FALLBACK_KEY: ${PACKAGE}-${CI_JOB_NAME}-${CI_DEFAULT_BRANCH}
+  MONOREPO_CACHE_FOLDER: ${CI_PROJECT_DIR}/.cache/monorepo
+  #
+  GIT_STRATEGY: fetch
+  GIT_DEPTH: 3
+
+before_script:
+  - wget -qO ./.gitlab-ci/monorepo.sh https://raw.githubusercontent.com/SocialGouv/gitlab-ci-yml/master/monorepo.sh
+  - sh ./.gitlab-ci/monorepo.sh skip && exit 0 || true
+  - cd "${PACKAGE_CWD}"
+
+Install:
+  stage: build
+  image: node:14.16-alpine3.13
+  needs: []
+  cache:
+    key:
+      files:
+        - yarn.lock
+      prefix: ${PACKAGE}-${CI_JOB_NAME}
+    paths:
+      - .cache
+      - ${PACKAGE_CWD}/node_modules
+      - node_modules
+  before_script:
+    - wget -qO ./.gitlab-ci/monorepo.sh https://raw.githubusercontent.com/SocialGouv/gitlab-ci-yml/master/monorepo.sh
+    - sh ./.gitlab-ci/monorepo.sh has-changed
+        --find-ignore "*/node_modules/*"
+        --verbose
+        ${PACKAGE_CWD} || true
+    - |
+      sh ./.gitlab-ci/monorepo.sh skip && {
+        rm -rf .cache/yarn "${PACKAGE_CWD}/node_modules" node_modules;
+        exit 0;
+      } || true
+  script:
+    - yarn config set cache-folder ${CI_PROJECT_DIR}/.cache/yarn
+    - yarn global add @socialgouv/yarn-workspace-focus-install --cwd ${PACKAGE_CWD}
+    - yarn exec yarn-workspace-focus-install -- --cwd ${PACKAGE_CWD}
+  artifacts:
+    expire_in: 1 week
+    paths:
+      - ${MONOREPO_CACHE_FOLDER}/skip
+      - ${PACKAGE_CWD}/node_modules
+      - node_modules
+
+lib logger:
+  stage: build
+  image: node:14.16-alpine3.13
+  needs: []
+  cache:
+    key:
+      files:
+        - yarn.lock
+      prefix: logger-Build
+    paths:
+      - infra/logger/lib
+    policy: pull
+  before_script:
+    - cd ${CI_PROJECT_DIR}
+  script:
+    - md5sum infra/logger/lib/*
+  artifacts:
+    expire_in: 1 week
+    paths:
+      - infra/logger/lib
+
+
+#
+#
+#
+
+Build:
+  stage: test
+  image: node:14.16-alpine3.13
+  cache:
+    key:
+      files:
+        - yarn.lock
+      prefix: ${PACKAGE}-${CI_JOB_NAME}
+    paths:
+      - ${PACKAGE_CWD}/lib
+  variables:
+    CI_DEBUG_TRACE: "true"
+  needs:
+    - job: Install
+      artifacts: true
+    - job: lib logger
+      artifacts: true
+  script:
+    # NOTE(douglasduteil): ensure to have clean build
+    - rm -rf lib
+    - yarn build
+  artifacts:
+    expire_in: 1 week
+    paths:
+      - ${PACKAGE_CWD}/lib
+
+#
+#
+#
+
+Lint:
+  stage: test
+  image: node:14.16-alpine3.13
+  needs:
+    - job: Install
+      artifacts: true
+    - job: lib logger
+      artifacts: true
+  script:
+    - yarn lint
+
+#
+#
+#
+
+Test:
+  stage: test
+  image: node:14.16-alpine3.13
+  needs:
+    - job: Install
+      artifacts: true
+    - job: lib logger
+      artifacts: true
+  script:
+    - yarn test
