+++ packages/code-du-travail-data/package.json
@@ -26,7 +26,7 @@
     "ingest": "node -r esm indexing/ingest",
     "precommit": "lint-staged",
     "prepush": "yarn test --bail --changedSince=master",
-    "test": "BUFFER_SIZE=10 SUGGEST_FILE=./indexing/__tests__/suggestion_data_test.txt jest"
+    "test": "jest"
   },
   "repository": {
     "type": "git",
